% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accion_atributos.R
\name{accion_atributos}
\alias{accion_atributos}
\title{Refuerza el envio de atributos cuando estos son multiples}
\usage{
accion_atributos(nombre_PLC, num_climas)
}
\arguments{
\item{nombre_PLC, }{num_climas}
}
\value{
json
}
\description{
Refuerza el envio de atributos cuando estos son multiples
}
\examples{
 accion_atributos("PLC_LOPY_RC1_7",3)

}
