% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_climas_auto_man.R
\name{control_climas_auto_man}
\alias{control_climas_auto_man}
\title{Gestiona auto manual climas}
\usage{
control_climas_auto_man(nombre_PLC)
}
\arguments{
\item{nombre_PLC}{}
}
\value{
json
}
\description{
Gestiona auto manual climas
}
\examples{
 control_climas_fallo_sondas("PLC P7")

}
