% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_climas_fallo_electro.R
\name{control_climas_fallo_electro}
\alias{control_climas_fallo_electro}
\title{Gestiona las climatizadoras frente a un problema en las electrovalvulas}
\usage{
control_climas_fallo_electro(nombre_PLC, num_climas)
}
\arguments{
\item{nombre_PLC, }{num_climas}
}
\value{
json
}
\description{
Gestiona las climatizadoras frente a un problema en las electrovalvulas
}
\examples{
 control_climas_fallo_electro("PLC P7",3)

}
