% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_climas_fallo_electro_horario.R
\name{control_climas_fallo_electro_horario}
\alias{control_climas_fallo_electro_horario}
\title{Gestiona las ELECTROVALVULAS cuando por horario deben de estar cerradas}
\usage{
control_climas_fallo_electro_horario(nombre_PLC, num_climas)
}
\arguments{
\item{nombre_PLC, }{num_climas}
}
\value{
json
}
\description{
Gestiona las ELECTROVALVULAS cuando por horario deben de estar cerradas
}
\examples{
 control_climas_fallo_electro_horario("PLC P7",3)

}
