% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_climas_fallo_sondas.R
\name{control_climas_fallo_sondas}
\alias{control_climas_fallo_sondas}
\title{Gestiona las climatizadoras frente a un problema en las sondas de Tº}
\usage{
control_climas_fallo_sondas(nombre_PLC, num_climas)
}
\arguments{
\item{nombre_PLC, }{num_climas}
}
\value{
json
}
\description{
Gestiona las climatizadoras frente a un problema en las sondas de Tº
}
\examples{
 control_climas_fallo_sondas("PLC P7",3)

}
