% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_climas_temperatura.R
\name{control_climas_temperatura}
\alias{control_climas_temperatura}
\title{Gestiona la climatizaci칩n (climas y electrov치lvulas) en base a temperatura de ERS}
\usage{
control_climas_temperatura(nombre_PLC, num_climas)
}
\arguments{
\item{nombre_PLC, }{num_climas}
}
\value{
json
}
\description{
Gestiona la climatizaci칩n (climas y electrov치lvulas) en base a temperatura de ERS
}
\examples{
 control_climas_temperatura("PLC P7",3)

}
