% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replica_datos.R
\name{replica_datos}
\alias{replica_datos}
\title{Replica datos}
\usage{
replica_datos(nombre_PLC, num_climas)
}
\arguments{
\item{nombre_PLC}{}
}
\value{
json
}
\description{
Replica datos
}
\examples{
 replica_datos("PLC P6_1")

}
